consolidate add commands to create namespace
