+++ src/main/java/sc/iview/create/AddVolume.java
@@ -1,4 +1,4 @@
-package sc.iview;
+package sc.iview.create;
 
 import net.imagej.Dataset;
 import net.imagej.ops.OpService;
@@ -6,6 +6,7 @@ import org.scijava.command.Command;
 import org.scijava.log.LogService;
 import org.scijava.plugin.Parameter;
 import org.scijava.plugin.Plugin;
+import sc.iview.SciView;
 
 /**
  * Created by kharrington on 6/21/17.
